web/requirements.txt
Babel==2.6.0
blinker==1.4
certifi==2019.3.9
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.9
MarkupSafe==1.1.1
numpy==1.16.3
pandas==0.24.2
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.1
six==1.12.0
visitor==0.1.3
Werkzeug==0.15.2
WTForms==2.2.1
